requirements.txt
mysql-connector-python
flask
gunicorn
datetime